Accept a YAML file, not an SVD file, when patching
A UI nitpick I noticed while testing #118. The Rust-based svdtools patch
help output suggests that the user provides an SVD file:

  $ svdtools patch --help
  svdtools-patch
  Patches an SVD file as specified by a YAML file

  USAGE:
      svdtools patch <SVD_FILE>

  ARGS:
      <SVD_FILE>    Path to input SVD file

  OPTIONS:
      -h, --help    Print help information

If we actually provide an SVD file, the tool errors.

  $ svdtools patch ../ral/svd/imxrt1176_cm7.svd
  [2022-08-19T12:47:48Z ERROR svdtools::cli] mapping values are not allowed in this context at line 11 column 24

Compare this with the Python documentation, which suggests that the user
provides a YAML file, not an SVD file.

  svd patch --help
  Usage: svd patch [OPTIONS] YAML_FILE

    Patches an SVD file as specified by a YAML file

  Options:
    --help  Show this message and exit.