Add a new test_utils::test_expected() helper function
This helper function makes it easy to write tests verifying that a patch
produces expected results.  In a test subdirectory you can now check in
a `patch.yaml` file and an `expected.svd`, and then write a 1-line test
to call `test_expected()` for this subdirectory.

If the patch results do not match the expected contents, the error
message will contain a unified diff of the mismatching contents.

I included a simple test of the clustering behavior as an example.