Make CLI optional (for library usage)
This is made by adding a "bin" feature that's enabled by default,
inspired by what was done in
https://github.com/rust-embedded/svd2rust/pull/619.

Additionally, support for JSON and YAML configs has also been made
optional but default. This includes what should be the only behavior
change: if neither were enabled at build-time, passing a config file is
an error.

The motivation came from https://github.com/Rahix/avr-device/issues/201,
but this should be a generally beneficial change for everybody.

Since some of the features of the CLI aren't very useful when using
svdtools as a library, they just vanish without the "bin" feature. This
is for two reasons: one is to reduce diff size, and two is to avoid
having to refactor all capabilities to have a library equivalent. There
is little disadvantage to this, for instance in avr-device we only use
the patching functionality (which is kept), and there's little prospect
of needing more.